+++ client-sdk/ts-web/rt/src/event.ts
@@ -1,6 +1,6 @@
 import * as oasis from '@oasisprotocol/client';
 
-export function toTag(module: string, code: number) {
+export function toKey(module: string, code: number) {
     const codeBuf = new ArrayBuffer(4);
     const codeDV = new DataView(codeBuf);
     codeDV.setUint32(0, code, false);