+++ client-sdk/ts-web/rt/docs/changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased changes
+## v0.2.0-alpha1
 
 Spotlight change: